Specify --without-icu in headupdate.sh for building HEAD of PostgreSQL.
The latest version of PostgreSQL has enabled the ICU support by default
for the build process. However, if the module for ICU is not installed,
the build process fails. This commit fixes this issue by updating
headupdate.sh to specify the --without-icu option while running
pgmake.sh to build PostgreSQL.